rename client -> wallet
