+++ src/oid4vci.rs
@@ -88,7 +88,7 @@
 //! [OpenID Connect]: (https://openid.net/specs/openid-connect-core-1_0.html)
 //! [RFC6749]: (https://www.rfc-editor.org/rfc/rfc6749.html)
 
-pub mod client;
+pub mod wallet;
 mod endpoint;
 pub mod pkce;
 pub mod provider;
