+++ src/oid4vci/wallet.rs
@@ -1,4 +1,4 @@
-//! # Client
+//! # Wallet
 
 mod authorization;
 mod credential;
